pyproject.toml
[tool.poetry]
name = "steganography"
version = "0.1.0"
description = "Simple script to hide text inside images."
authors = ["Jefferson Bizarri <jefferson.bizarri@gmail.com>"]
license = "MIT License"

[tool.poetry.dependencies]
python = "~3.8"
Pillow = "8.1.2"


[tool.poetry.dev-dependencies]
black = "20.8b1"
pytest = "6.2.2"
coverage = "5.5"
tox = "3.23.0"
flake8 = "~3.8"
pdoc3 = "0.9.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
